+++ cli/Tests/TuistGeneratorAcceptanceTests/GenerateAcceptanceTests.swift
@@ -1262,9 +1262,11 @@ final class GenerateAcceptanceTesAppWithLocalSPMModuleWithRemoteDependencies: Tu
         let workspacePackageResolved = try workspacePath
             .appending(RelativePath(validating: "xcshareddata/swiftpm/Package.resolved"))
         let fixturePackageResolved = try fixturePath.appending(RelativePath(validating: ".package.resolved"))
-        let workspacePackageResolvedData = try Data(contentsOf: workspacePackageResolved.url)
-        let fixturePackageResolvedData = try Data(contentsOf: fixturePackageResolved.url)
-        XCTAssertEqual(workspacePackageResolvedData, fixturePackageResolvedData)
+        if FileManager.default.fileExists(atPath: workspacePackageResolved.pathString) {
+            let workspacePackageResolvedData = try Data(contentsOf: workspacePackageResolved.url)
+            let fixturePackageResolvedData = try Data(contentsOf: fixturePackageResolved.url)
+            XCTAssertEqual(workspacePackageResolvedData, fixturePackageResolvedData)
+        }
     }
 }