+++ server/lib/tuist/application.ex
@@ -4,6 +4,7 @@ defmodule Tuist.Application do
   use Application
   use Boundary, top_level?: true, deps: [Tuist, TuistWeb]
 
+  alias Tuist.Builds.Build
   alias Tuist.CommandEvents
   alias Tuist.DBConnection.TelemetryListener
   alias Tuist.Environment
@@ -86,6 +87,7 @@ defmodule Tuist.Application do
         storage: :memory,
         labels: %{
           app: {:static, "tuist-server"},
+          service_name: {:static, "tuist-server"},
           env: {:static, to_string(Environment.env())},
           level: :level
         },
@@ -102,6 +104,7 @@ defmodule Tuist.Application do
         Tuist.ClickHouseRepo,
         Tuist.IngestRepo,
         Supervisor.child_spec(CommandEvents.Buffer, id: CommandEvents.Buffer),
+        Supervisor.child_spec(Build.Buffer, id: Build.Buffer),
         Supervisor.child_spec(Logs.Buffer, id: Logs.Buffer),
         Supervisor.child_spec(XcodeGraph.Buffer, id: XcodeGraph.Buffer),
         Supervisor.child_spec(XcodeProject.Buffer, id: XcodeProject.Buffer),