+++ server/lib/tuist/repo.ex
@@ -4,21 +4,6 @@ defmodule Tuist.Repo do
     adapter: Ecto.Adapters.Postgres,
     pool_timeout: 15_000
 
-  def timescale_available? do
-    query = """
-    SELECT EXISTS (
-      SELECT 1
-      FROM pg_extension
-      WHERE extname = 'timescaledb'
-    )
-    """
-
-    case Ecto.Adapters.SQL.query!(__MODULE__, query, []) do
-      %{rows: [[true]]} -> true
-      %{rows: [[false]]} -> false
-    end
-  end
-
   def running? do
     Enum.member?(Ecto.Repo.all_running(), __MODULE__)
   end