+++ server/lib/tuist_web/components/app_layout_components.ex
@@ -325,6 +325,8 @@ defmodule TuistWeb.AppLayoutComponents do
           label={breadcrumb.label}
           show_avatar={Map.get(breadcrumb, :show_avatar, false)}
           avatar_color={Map.get(breadcrumb, :avatar_color)}
+          badge_label={breadcrumb[:badge] && breadcrumb.badge.label}
+          badge_color={breadcrumb[:badge] && breadcrumb.badge.color}
         >
           <:icon :if={Map.get(breadcrumb, :icon)}><.icon name={Map.get(breadcrumb, :icon)} /></:icon>
           <.breadcrumb_item
@@ -337,6 +339,8 @@ defmodule TuistWeb.AppLayoutComponents do
             show_avatar={Map.get(breadcrumb_item, :show_avatar, false)}
             avatar_color={Map.get(breadcrumb_item, :avatar_color)}
             icon={Map.get(breadcrumb_item, :icon)}
+            badge_label={breadcrumb_item[:badge] && breadcrumb_item.badge.label}
+            badge_color={breadcrumb_item[:badge] && breadcrumb_item.badge.color}
           />
         </.breadcrumb>
       <% end %>